Move navbar logic from app to separate comp file - MainNavbar
